renamed the >model>user to slot and add instructor.
