+++ controllers/admin.js
@@ -1,5 +1,5 @@
 const jwt = require("jsonwebtoken");
-const User = require('../models/User');
+const User = require('../models/Slot');
 // const middleware = require('../middleware/authorization'); 
 
 exports.getUsers = (req, res) => {
