+++ controllers/slot.js
@@ -1,4 +1,4 @@
-const User = require('../models/SlotSchema');
+const Slot = require('../models/Slot');
 
 //checking
 exports.addSlot = async (req,res) => {
@@ -6,16 +6,10 @@ exports.addSlot = async (req,res) => {
         || !req.body.status|| !req.body.booking) {
     return res.status(400).json({msg: "Invalid Data"})
     }
-    const user = await User.findById(req.query.body)
+    const user = await Slot.findById(req.query.body)
     if(user){
-        let user = new User ({
-            date : req.body.date,
-            time : req.body.time,
-            clientLimit : req.body.clientLimit , 
-            instructor : req.body.instructor , 
-            instructorName : req.body.instructorName,
-            status : req.body.status, 
-            booking : req.body.booking
+        let user = new Slot ({
+            ...req.body,
         });
         user.save() 
         .then(b=>{
@@ -27,12 +21,12 @@ exports.addSlot = async (req,res) => {
             res.status(400).json({err});
         })
     } else {
-        res.status(400).json({msg: "Category does not exist"});
+        res.status(400).json({msg: "Instructor does not exist"});
     }  
 }    
 
 exports.modifySlot =  ( req , res ) => {
-    User.findById(req.query.id)
+    Slot.findById(req.query.id)
     .then(b=>{
         if (b) {
             b.date = req.body.date,
@@ -54,7 +48,7 @@ exports.modifySlot =  ( req , res ) => {
 }
 
 exports.getSlots =  ( req , res ) => {
-    User.find({})
+    Slot.find({})
     .then(b => {
         return res.status(200).json({ b: b })
     })
@@ -68,7 +62,7 @@ exports.getSlotById =  ( req , res ) => {
     if (!req.query.id) {
         return res.status(400).json({ msg: "You need to send the ID!" })
     }
-    User.find({ _id: req.query.id })
+    Slot.find({ _id: req.query.id })
     .then(b => {
         return res.status(200).json({ b: b });
     })
@@ -80,7 +74,7 @@ exports.searchSlotByDate =  ( req , res ) => {
     if (!req.query.date) {
         return res.status(400).json({ msg: "Please enter date!" })
     }
-    User.find({ date: req.query.date })
+    Slot.find({ date: req.query.date })
     .then(b => {
         return res.status(200).json({ b: b });
     })
@@ -90,7 +84,7 @@ exports.searchSlotByDate =  ( req , res ) => {
 }
 
 exports.deleteSlot =  ( req , res ) => {
-    User.findByIdAndDelete({_id: req.query.id })
+    Slot.findByIdAndDelete({_id: req.query.id })
     .then(b => {
         return res.status(200).json({ b: b });
     })
