+++ models/Slot.js
@@ -1,7 +1,6 @@
 const mongoose = require("mongoose");
 
 const ClientSchema = mongoose.Schema({
-
     fullName: {
       type: String,
       required: true,
